refactor: centralize spherical coord helpers
